Make prepareInLoadThread async, return a Future.
